+++ .circleci/config.yml
@@ -3,7 +3,7 @@ version: 2.1
 jobs:
   build-service:
     docker:
-      - image: alexfalkowski/go:1.10
+      - image: alexfalkowski/go:1.12
     working_directory: ~/web
     steps:
       - checkout
@@ -53,7 +53,7 @@ jobs:
     resource_class: large
   build-docker:
     docker:
-      - image: alexfalkowski/go:1.10
+      - image: alexfalkowski/go:1.12
     working_directory: ~/web
     steps:
       - checkout
@@ -76,7 +76,7 @@ jobs:
     resource_class: large
   push-docker:
     docker:
-      - image: alexfalkowski/go:1.10
+      - image: alexfalkowski/go:1.12
     working_directory: ~/web
     steps:
       - checkout
