+++ .golangci.yml
@@ -7,7 +7,6 @@ linters-settings:
 linters:
   enable-all: true
   disable:
-    - exportloopref
     - tagalign
     - bodyclose
     - depguard
@@ -15,3 +14,4 @@ linters:
     - paralleltest
     - ireturn
     - wrapcheck
+    - tenv
