+++ go.mod
@@ -6,7 +6,7 @@ toolchain go1.23.6
 
 require (
 	github.com/alexfalkowski/go-health v1.21.0
-	github.com/alexfalkowski/go-service v1.427.0
+	github.com/alexfalkowski/go-service v1.428.0
 	go.uber.org/fx v1.23.0
 )
