+++ internal/health/registrations.go
