+++ internal/site/books/books.yaml
