+++ internal/site/books/model.go
